workspace-template/requirements.txt
a2a-python>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.3.0
langfuse>=2.0.0
httpx>=0.27.0
uvicorn>=0.30.0
pyyaml>=6.0
starlette>=0.38.0
